+++ docs/changes.rst
@@ -28,6 +28,8 @@ Released: not yet
 
 **Enhancements:**
 
+* Improved logging in WBEM listener and its test module.
+
 **Cleanup:**
 
 **Known issues:**